Move related genes module to new "packs" folder
