+++ src/js/ideogram.js
@@ -56,7 +56,7 @@ import {
   drawChromosomeLabels, rotateChromosomeLabels
 } from './views/chromosome-labels.js';
 
-import {_initRelatedGenes, plotRelatedGenes} from './related-genes';
+import {_initRelatedGenes, plotRelatedGenes} from './packs/related-genes';
 
 export default class Ideogram {
   constructor(config) {
